Cached shifters and new symmetry implementation
