+++ scarlet/interpolation.py
@@ -1,4 +1,5 @@
 import numpy as np
+from .cache import Cache
 
 
 def get_projection_slices(image, shape, yx0=None):
@@ -296,6 +297,42 @@ def get_separable_kernel(dy, dx, kernel=lanczos, **kwargs):
     kyx = np.outer(ky, kx)
     return kyx, y_window, x_window
 
+
+def mk_shifter(shape):
+    ''' Performs shifts in the Fourier domain on Fourier objects
+
+    Parameters:
+    -----------
+    shape: array
+        shape of the 2-D array to shift
+    Returns:
+    --------
+    result: Fourier
+        A Fourier object with shifted arrays
+    '''
+
+    #ssert all([s is fftpack.helper.next_fast_len(s) for s in fft_shape])
+
+    #Name of the chached shifts. Here I deviate from convention voluntarilly because several different methods
+    # might cache useful shifts.
+    name = 'fft_shift'
+    key = (shape[0], shape[1])
+    try:
+        shift_y, shift_x = Cache.check(name, key)
+    except KeyError:
+
+        freq_x = np.fft.rfftfreq(shape[1])
+        freq_y = np.fft.fftfreq(shape[0])
+
+        # Shift the signal to recenter it, negative because math is opposite from
+        # pixel direction
+        shift_y = np.exp(-1j * 2 * np.pi * freq_y)
+        shift_x = np.exp(-1j * 2 * np.pi * freq_x)
+
+    shifters = (shift_y, shift_x)
+    Cache.set(name, key, shifters)
+
+    return shifters
 
 def sinc_interp(coord_hr, coord_lr, sample_lr):
     '''
